warning fixes, stable

+++ SDLProject/include/Character.hpp
@@ -9,6 +9,6 @@ public:
 	inline Character(Properties* props) : GameObject(props) {}
 
 	virtual void Draw() = 0;
-	virtual void Update(double dt) = 0;
+	virtual void Update(Uint64 dt) = 0;
 	virtual void Clean() = 0;
 };
